refactor: Move helper functions to the states.rs
