+++ lightning/src/ln/peers/handshake/mod.rs
@@ -2,16 +2,10 @@
 //! Handshake states can be advanced automatically, or by manually calling the appropriate step.
 //! Once complete, returns an instance of Conduit.
 
-use bitcoin::secp256k1;
-
-use bitcoin::hashes::{Hash, HashEngine};
-use bitcoin::hashes::sha256::Hash as Sha256;
 use bitcoin::secp256k1::{PublicKey, SecretKey};
 
-use ln::peers::{chacha, hkdf};
-use ln::peers::conduit::{Conduit, SymmetricKey};
+use ln::peers::conduit::Conduit;
 use ln::peers::handshake::acts::Act;
-use ln::peers::handshake::hash::HandshakeHash;
 use ln::peers::handshake::states::{HandshakeState2, UninitiatedHandshakeState, AwaitingActOneHandshakeState};
 
 mod acts;
@@ -48,23 +42,6 @@ impl PeerHandshake {
 		self.remote_public_key
 	}
 
-	fn initialize_state(public_key: &PublicKey) -> (HandshakeHash, [u8; 32]) {
-		let protocol_name = b"Noise_XK_secp256k1_ChaChaPoly_SHA256";
-		let prologue = b"lightning";
-
-		let mut sha = Sha256::engine();
-		sha.input(protocol_name);
-		let chaining_key = Sha256::from_engine(sha).into_inner();
-
-		let mut initial_hash_preimage = chaining_key.to_vec();
-		initial_hash_preimage.extend_from_slice(prologue.as_ref());
-
-		let mut hash = HandshakeHash::new(initial_hash_preimage.as_slice());
-		hash.update(&public_key.serialize());
-
-		(hash, chaining_key)
-	}
-
 	/// Process act dynamically
 	/// # Arguments
 	/// `input`: Byte slice received from peer as part of the handshake protocol
@@ -91,79 +68,6 @@ impl PeerHandshake {
 		self.state = Some(next_state);
 
 		result
-	}
-
-	fn calculate_act_message(local_private_key: &SecretKey, remote_public_key: &PublicKey, chaining_key: [u8; 32], hash: &mut HandshakeHash) -> ([u8; 50], SymmetricKey, SymmetricKey) {
-		let local_public_key = Self::private_key_to_public_key(local_private_key);
-
-		hash.update(&local_public_key.serialize());
-
-		let ecdh = Self::ecdh(local_private_key, &remote_public_key);
-		let (chaining_key, temporary_key) = hkdf::derive(&chaining_key, &ecdh);
-		let tagged_ciphertext = chacha::encrypt(&temporary_key, 0, &hash.value, &[0; 0]);
-
-		hash.update(&tagged_ciphertext);
-
-		let mut act = [0u8; 50];
-		act[1..34].copy_from_slice(&local_public_key.serialize());
-		act[34..].copy_from_slice(tagged_ciphertext.as_slice());
-
-		(act, chaining_key, temporary_key)
-	}
-
-	// Due to the very high similarity of acts 1 and 2, this method is used to process both
-	fn process_act_message(act_bytes: [u8; 50], local_private_key: &SecretKey, chaining_key: SymmetricKey, hash: &mut HandshakeHash) -> Result<(PublicKey, SymmetricKey, SymmetricKey), String> {
-		let version = act_bytes[0];
-		if version != 0 {
-			// this should not crash the process, hence no panic
-			return Err("unexpected version".to_string());
-		}
-
-		let mut ephemeral_public_key_bytes = [0u8; 33];
-		ephemeral_public_key_bytes.copy_from_slice(&act_bytes[1..34]);
-		let ephemeral_public_key = if let Ok(public_key) = PublicKey::from_slice(&ephemeral_public_key_bytes) {
-			public_key
-		} else {
-			return Err("invalid remote ephemeral public key".to_string());
-		};
-
-		let mut chacha_tag = [0u8; 16];
-		chacha_tag.copy_from_slice(&act_bytes[34..50]);
-
-		// process the act message
-
-		// update hash with partner's pubkey
-		hash.update(&ephemeral_public_key.serialize());
-
-		// calculate ECDH with partner's pubkey and local privkey
-		let ecdh = Self::ecdh(local_private_key, &ephemeral_public_key);
-
-		// HKDF(chaining key, ECDH) -> chaining key' + next temporary key
-		let (chaining_key, temporary_key) = hkdf::derive(&chaining_key, &ecdh);
-
-		// Validate chacha tag (temporary key, 0, hash, chacha_tag)
-		let _tag_check = chacha::decrypt(&temporary_key, 0, &hash.value, &chacha_tag)?;
-
-		hash.update(&chacha_tag);
-
-		Ok((ephemeral_public_key, chaining_key, temporary_key))
-	}
-
-	fn private_key_to_public_key(private_key: &SecretKey) -> PublicKey {
-		let curve = secp256k1::Secp256k1::new();
-		let pk_object = PublicKey::from_secret_key(&curve, &private_key);
-		pk_object
-	}
-
-	fn ecdh(private_key: &SecretKey, public_key: &PublicKey) -> SymmetricKey {
-		let curve = secp256k1::Secp256k1::new();
-		let mut pk_object = public_key.clone();
-		pk_object.mul_assign(&curve, &private_key[..]).expect("invalid multiplication");
-
-		let preimage = pk_object.serialize();
-		let mut sha = Sha256::engine();
-		sha.input(preimage.as_ref());
-		Sha256::from_engine(sha).into_inner()
 	}
 }
